+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 *   Add a bash auto-complete tool [#2](https://github.com/TylerTemp/docpie/issues/2)
 *   Document needs a better organization
 
+## 0.3.3
+
+*   [fix] a logging is not use `docpie` logger
+
 ## 0.3.2
 
 *   [fix] options section titles were not set appropriate when options sections's